feat: :pencil2: convert some values to be in `dp` instead of `px`
convert the values of `startLocationX`, `startLocationY`, and callbacks parameters `x`, `y` to be in dp instead of px

`startLocationX` and `startLocationY` would change because they are mesured in `dp` now instead of `px` as the Flutter convention

+++ android/src/main/kotlin/dev/moaz/dash_bubble/src/BroadcastListener.kt
@@ -41,10 +41,10 @@ class BroadcastListener(channel: MethodChannel) : BroadcastReceiver() {
     }
 
     /** This method is used to get the coordinates from the intent extras. */
-    private fun getCoordinatesValues(intent: Intent): HashMap<String, Float> {
+    private fun getCoordinatesValues(intent: Intent): HashMap<String, Double> {
         return hashMapOf(
-            Constants.X_AXIS_VALUE to intent.getFloatExtra(Constants.X_AXIS_VALUE, 0f),
-            Constants.Y_AXIS_VALUE to intent.getFloatExtra(Constants.Y_AXIS_VALUE, 0f)
+            Constants.X_AXIS_VALUE to intent.getDoubleExtra(Constants.X_AXIS_VALUE, 0.0),
+            Constants.Y_AXIS_VALUE to intent.getDoubleExtra(Constants.Y_AXIS_VALUE, 0.0)
         )
     }
 }
